% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_by_def}
\alias{bin_by_def}
\title{Binning by definition (bounds and centers)}
\usage{
bin_by_def(new_data_vec, bin_definition, output = "centers")
}
\arguments{
\item{new_data_vec}{data vector to bin using defined partion.}

\item{bin_definition}{list containing bounds of binning partition.}

\item{output}{"centers" or "all" to select if function will return only bin centers vector or list with bin centers and defining bounds, respectively.}
}
\value{
returns either the new binned data, or a list of the data and defining bounds
}
\description{
univariate binning based on bins definition (potentially developed from different data source, i.e. train bins then bin test)
define bin boundaries, adapt outermost for the range of data to be binned, then use .bincode to drop to primitive to relabel
}

