% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_loss.R
\name{bin_loss}
\alias{bin_loss}
\title{Creating bin definition lists from vectors of bounds and centers}
\usage{
bin_loss(reduced_data, type = "average")
}
\arguments{
\item{reduced_data}{Reduced binned data from bivariate binning function}

\item{type}{Type of loss summary "average" or "total"}
}
\value{
Vector of loss summaries
}
\description{
Take in reduced-binned data, output loss summaries
}
\examples{
reduced_data <- rect_bin_2d(data=diamonds, xcol="carat", ycol="price",originx=0,originy=0,widthx=.25,widthy=100, output="reduced")
bin_loss(reduced_data)
bin_loss(reduced_data, type="total")
}

