% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_scatterplots.R
\name{binned_scatter}
\alias{binned_scatter}
\title{Creating bin definition lists from vectors of bounds and centers}
\usage{
binned_scatter(reduced_data, count_scale = "identity", freq_binned = FALSE)
}
\arguments{
\item{reduced_data}{Reduced binned data from either \code{rect_bin_2d} or \code{rand_rect_bin_2d}}

\item{count_scale}{Transformation for shading scale for bin counts}

\item{freq_binned}{TRUE/FALSE does reduced binned data has been frequency binned.}
}
\value{
Vector of loss summaries
}
\description{
Construct binned scatterplot from reduced binned data, build upon \code{geom_tile} from \code{ggplot2} 
Takes in either reduced-binned data with raw counts or with frequency-binned counts. Allow for raw/log/log10 scaling
}
\examples{
n_freq <-5
red_data <- rect_bin_2d(data=diamonds, xcol="carat",ycol="price",originx=0,originy=0,widthx=.25,widthy=1000, output="reduced")
red_freq_data <- freq_bin(red_data,  bin_type="standard", count_type="log10", n_freq=n_freq, pretty=FALSE, freq_col="freq")
binned_scatter(red_freq_data, count_scale="log10", freq_binned = FALSE)
binned_scatter(red_freq_data, count_scale="log10",freq_binned = TRUE) +
 xlab("Carat Weight") + ylab("Price (dollars)") +
 ggtitle("Binned Scatterplot of Diamond Price By Carat Weight")
}

