% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_scatterplots.R
\name{binned_scatter_raw}
\alias{binned_scatter_raw}
\title{Creating bin definition lists from vectors of bounds and centers}
\usage{
binned_scatter_raw(data, xcol, ycol, originx, originy, widthx, widthy,
  count_type = "raw", freq_binned = FALSE, freq_bin_type = "standard",
  n_freq = 1, pretty = FALSE)
}
\arguments{
\item{data}{data frame with at least two numeric columns to be binned}

\item{xcol}{name of column to bin for x dimension}

\item{ycol}{name of column to bin for y dimension}

\item{originx}{origin numeric value at which to begin standard rectangular binning for x dimension. This should be selected at or below minimum value in \code{xs}.}

\item{originy}{origin numeric value at which to begin standard rectangular binning for y dimension. This should be selected at or below minimum value in \code{xs}.}

\item{widthx}{Positive numeric value at which to space standard rectangular bins for x dimension.}

\item{widthy}{Positive numeric value at which to space standard rectangular bins for y dimension.}

\item{count_type}{"raw","log", or "log10" transformation of counts specified}

\item{freq_binned}{TRUE/FALSE for using frequency binning for discretized color scale.}

\item{freq_bin_type}{"standard" or "quantile" binning specification}

\item{n_freq}{positive integer specifying number of bins}

\item{pretty}{Logical for using pretty function for cleaner frequency bin breaks}
}
\value{
Vector of loss summaries
}
\description{
All-in-one Binning and Scatterplot Creating Super Function
Need to specify all origins, widths, transformations, frequency bins
}
\examples{
binned_scatter_raw(data=diamonds, xcol="carat", ycol="price", originx=0,
originy=0, widthx=.25, widthy=1000, count_type = "log10")
binned_scatter_raw(data=diamonds, xcol="carat", ycol="price", originx=0,
                   originy=0, widthx=.25, widthy=1000, count_type="raw",
                   freq_binned=TRUE, freq_bin_type="quantile", n_freq=5, pretty=FALSE)
}

