% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{freq_bin}
\alias{freq_bin}
\title{Frequency binning of reduced binned data}
\usage{
freq_bin(reduced_data, bin_type = "standard", count_type = "raw",
  n_freq = 5, pretty = FALSE, freq_col = "freq")
}
\arguments{
\item{reduced_data}{Requires reduced binned data output from \code{rect_bin_2d} function, number of freq breaks and type of freq binning}

\item{bin_type}{"standard" or "quantile" binning specification}

\item{count_type}{"raw","log", or "log10" transformation of counts specified}

\item{n_freq}{positive integer specifying number of bins}

\item{pretty}{Logical for using pretty function for cleaner frequency bin breaks}

\item{freq_col}{name of frequency column to be binned}
}
\value{
output of frequency bin values, labels and loss are attached the original then returned
}
\description{
allows for standard or quantile binning of bin counts that may be raw,log,log10 (6 combinations)
for use in discretizing binned scatterplot shading.
}
\examples{
reduced_data <- rect_bin_2d(data=diamonds, xcol="carat", ycol="price",originx=0,originy=0,widthx=.5,widthy=1000, output="reduced")
freq_red_data <- freq_bin(reduced_data, bin_type="standard", count_type="raw", n_freq=5, pretty=FALSE, freq_col="freq")
head(freq_red_data)
freq_log10_red_data <- freq_bin(reduced_data, bin_type="standard", count_type="log10", n_freq=5, pretty=FALSE, freq_col="freq")
head(freq_log10_red_data)
sum(freq_log10_red_data$log10_freq_loss)
}

