% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{gen_rect_bin_1d}
\alias{gen_rect_bin_1d}
\title{General rectangular binning}
\usage{
gen_rect_bin_1d(xs, bounds, output = "centers")
}
\arguments{
\item{xs}{Vector of numeric values to bin}

\item{bounds}{Vector of unique, increasing boundaries values that specify bin partitions}

\item{output}{Parameter reserved for expanding output types in future package versions.}
}
\value{
bin centers as a vector corresponding to \code{xs} vector
}
\description{
Generalized Rectangular 1d Binning : given the set of bin breaks
}
\examples{
xs <- runif(100,1,25)
bounds <- c(1,4,7,12,13,19,23,24,25,29)
gen_rect_bin_1d(xs, bounds ,output="centers")
}

