% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{make_bin_def}
\alias{make_bin_def}
\title{Creating bin definition lists from vectors of bounds and centers}
\usage{
make_bin_def(centers, bounds)
}
\arguments{
\item{centers}{vector of bin centers}

\item{bounds}{vector of bin boundaries}
}
\value{
Returns list of bin definition
}
\description{
Helper function to combine vectors of bin centers and bounds into a list to be used in other functions that bin by definition.
}
\examples{
make_bin_def(1:10,seq(.5,11.5,by=1))
}

