% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{quant_bin_1d}
\alias{quant_bin_1d}
\title{Quantile rectangular binning}
\usage{
quant_bin_1d(xs, nbin, output = "centers")
}
\arguments{
\item{xs}{Vector of numeric values to bin.}

\item{nbin}{Number of quantile based bins, must be a positive integer.}

\item{output}{Parameter reserved for expanding output types in future package versions.}
}
\value{
bin centers as a vector corresponding to \code{xs} vector
}
\description{
Quantile 1d Binning: used for binning the counts values by quantile, must define vector of counts and number of bins.
}
\examples{
quant_bin_1d(ggplot2::diamonds$price,4)
}

