% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{rect_bin_1d}
\alias{rect_bin_1d}
\title{Standard rectangular binning}
\usage{
rect_bin_1d(xs, origin, width, output = "centers")
}
\arguments{
\item{xs}{Vector of numeric values to bin.}

\item{origin}{numeric value at which to begin standard rectangular binning. This should be selected at or below minimum value in \code{xs}.}

\item{width}{Positive numeric value at which to space standard rectangular bins.}

\item{output}{Parameter reserved for expanding output types in future package versions.}
}
\value{
bin centers as a vector corresponding to \code{xs} vector
}
\description{
Standard Rectangular 1d Binning: given origin and consistent bin width.
}
\examples{
xs <- ggplot2::diamonds$price; nbins=4; origin=min(xs); width=diff(range(xs))/nbins
rect_bin_1d(rnorm(1000,0,1),origin=-4,width=1)
}

