% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{rect_bin_2d}
\alias{rect_bin_2d}
\title{Independant 2d binning grid}
\usage{
rect_bin_2d(data, xcol, ycol, originx, originy, widthx, widthy,
  output = "full")
}
\arguments{
\item{data}{data frame with at least two numeric columns to be binned}

\item{xcol}{name of column to bin for x dimension}

\item{ycol}{name of column to bin for y dimension}

\item{originx}{origin numeric value at which to begin standard rectangular binning for x dimension. This should be selected at or below minimum value in \code{xs}.}

\item{originy}{origin numeric value at which to begin standard rectangular binning for y dimension. This should be selected at or below minimum value in \code{xs}.}

\item{widthx}{Positive numeric value at which to space standard rectangular bins for x dimension.}

\item{widthy}{Positive numeric value at which to space standard rectangular bins for y dimension.}

\item{output}{"full" (default),"centers",or"reduced"}
}
\value{
returns data frame of specified type
}
\description{
Independent 2d Rectangular Binning for Binned Scatterplots. Built to return a data frame containing:
full data - the input data with binned xs and ys as new columns,
OR
centers data - the new binned xs and ys centers only,
OR
reduced binned data - one row per bivariate bin with columns for x and y centers, bin counts, and euclidean spatial losses per bin.
}
\examples{
full_data <- rect_bin_2d(data=diamonds, xcol="carat", ycol="price",originx=0,originy=0,widthx=.5,widthy=1000, output="full")
head(full_data)
centers_data <- rect_bin_2d(data=diamonds, xcol="carat", ycol="price",originx=0,originy=0,widthx=.5,widthy=1000, output="centers")
head(centers_data)
reduced_data <- rect_bin_2d(data=diamonds, xcol="carat", ycol="price",originx=0,originy=0,widthx=.5,widthy=1000, output="reduced")
head(reduced_data)
}

